Move transform_type header to cpp file
